+++ RELEASENOTES.md
@@ -2,6 +2,8 @@
 
 ### Next version
 - DEV adding Bluetooth3 OBD2 device
+- DEV ota
+- new message dialog
 
 ### v2.5.2 2021-11-19
 - MEB optimizations (cell screen, min/max voltage)